Make it so that modifying tests would be easier
