+++ frontend/src/embedding-sdk-bundle/components/public/dashboard/EditableDashboard/EditableDashboard.unit.spec.tsx
@@ -10,6 +10,7 @@ import { screen, waitFor, within } from "__support__/ui";
 import { ROOT_COLLECTION } from "metabase/entities/collections";
 import { createMockCollection } from "metabase-types/api/mocks";
 
+import { addSubscriptionTests } from "../StaticDashboard/sharedTests.spec";
 import {
   type SetupSdkDashboardOptions,
   setupSdkDashboard,
@@ -97,87 +98,7 @@ describe("EditableDashboard", () => {
     ).not.toBeInTheDocument();
   });
 
-  describe("Subscriptions Button", () => {
-    it.each([
-      {
-        isSlackConfigured: false,
-      },
-      {
-        isSlackConfigured: false,
-      },
-    ])(
-      "should show subscriptions button if subscriptions are enabled and email is set up (isSlackConfigured: $isSlackConfigured)",
-      async ({ isSlackConfigured }) => {
-        await setup({
-          props: { withSubscriptions: true },
-          isEmailConfigured: true,
-          isSlackConfigured,
-        });
-
-        await waitFor(() => {
-          expect(screen.getByTestId("dashboard-header")).toBeInTheDocument();
-        });
-
-        const dashboardHeader = within(screen.getByTestId("dashboard-header"));
-
-        expect(
-          await dashboardHeader.findByLabelText("Subscriptions"),
-        ).toBeInTheDocument();
-      },
-    );
-
-    it.each([
-      {
-        isEmailConfigured: false,
-        isSlackConfigured: false,
-        withSubscriptions: false,
-      },
-      {
-        isEmailConfigured: false,
-        isSlackConfigured: false,
-        withSubscriptions: true,
-      },
-      {
-        isEmailConfigured: false,
-        isSlackConfigured: true,
-        withSubscriptions: false,
-      },
-      {
-        isEmailConfigured: false,
-        isSlackConfigured: true,
-        withSubscriptions: true,
-      },
-      {
-        isEmailConfigured: true,
-        isSlackConfigured: false,
-        withSubscriptions: false,
-      },
-      {
-        isEmailConfigured: true,
-        isSlackConfigured: true,
-        withSubscriptions: false,
-      },
-    ])(
-      "should not show subscriptions button if subscriptions are disabled or email is not configured (isEmailConfigured: $isEmailConfigured, isSlackConfigured: $isSlackConfigured, withSubscriptions: $withSubscriptions)",
-      async ({ isEmailConfigured, isSlackConfigured, withSubscriptions }) => {
-        await setup({
-          props: { withSubscriptions },
-          isEmailConfigured,
-          isSlackConfigured,
-        });
-
-        await waitFor(() => {
-          expect(screen.getByTestId("dashboard-header")).toBeInTheDocument();
-        });
-
-        const dashboardHeader = within(screen.getByTestId("dashboard-header"));
-
-        expect(
-          dashboardHeader.queryByLabelText("Subscriptions"),
-        ).not.toBeInTheDocument();
-      },
-    );
-  });
+  addSubscriptionTests(setup);
 
   it("should allow to create a new question in addition to adding existing questions", async () => {
     await setup();
