+++ frontend/src/embedding-sdk-bundle/components/public/dashboard/StaticDashboard/StaticDashboard.unit.spec.tsx
@@ -8,6 +8,7 @@ import {
 } from "../tests/setup";
 
 import { StaticDashboard } from "./StaticDashboard";
+import { addSubscriptionTests } from "./sharedTests.spec";
 
 jest.mock("metabase/common/hooks/use-locale", () => ({
   useLocale: jest.fn(),
@@ -101,87 +102,7 @@ describe("StaticDashboard", () => {
     ).not.toBeInTheDocument();
   });
 
-  describe("Subscriptions Button", () => {
-    it.each([
-      {
-        isSlackConfigured: false,
-      },
-      {
-        isSlackConfigured: false,
-      },
-    ])(
-      "should show subscriptions button if subscriptions are enabled and email is set up (isSlackConfigured: $isSlackConfigured)",
-      async ({ isSlackConfigured }) => {
-        await setup({
-          props: { withSubscriptions: true },
-          isEmailConfigured: true,
-          isSlackConfigured,
-        });
-
-        await waitFor(() => {
-          expect(screen.getByTestId("dashboard-header")).toBeInTheDocument();
-        });
-
-        const dashboardHeader = within(screen.getByTestId("dashboard-header"));
-
-        expect(
-          await dashboardHeader.findByLabelText("Subscriptions"),
-        ).toBeInTheDocument();
-      },
-    );
-
-    it.each([
-      {
-        isEmailConfigured: false,
-        isSlackConfigured: false,
-        withSubscriptions: false,
-      },
-      {
-        isEmailConfigured: false,
-        isSlackConfigured: false,
-        withSubscriptions: true,
-      },
-      {
-        isEmailConfigured: false,
-        isSlackConfigured: true,
-        withSubscriptions: false,
-      },
-      {
-        isEmailConfigured: false,
-        isSlackConfigured: true,
-        withSubscriptions: true,
-      },
-      {
-        isEmailConfigured: true,
-        isSlackConfigured: false,
-        withSubscriptions: false,
-      },
-      {
-        isEmailConfigured: true,
-        isSlackConfigured: true,
-        withSubscriptions: false,
-      },
-    ])(
-      "should not show subscriptions button if subscriptions are disabled or email is not configured (isEmailConfigured: $isEmailConfigured, isSlackConfigured: $isSlackConfigured, withSubscriptions: $withSubscriptions)",
-      async ({ isEmailConfigured, isSlackConfigured, withSubscriptions }) => {
-        await setup({
-          props: { withSubscriptions },
-          isEmailConfigured,
-          isSlackConfigured,
-        });
-
-        await waitFor(() => {
-          expect(screen.getByTestId("dashboard-header")).toBeInTheDocument();
-        });
-
-        const dashboardHeader = within(screen.getByTestId("dashboard-header"));
-
-        expect(
-          dashboardHeader.queryByLabelText("Subscriptions"),
-        ).not.toBeInTheDocument();
-      },
-    );
-  });
+  addSubscriptionTests(setup);
 
   it("should not show subscriptions button if subscriptions are disabled", async () => {
     await setup({
